+++ packages/pyroscope-flamegraph/src/ProfilerTable.tsx
@@ -1,8 +1,14 @@
-import React, { useRef, RefObject } from 'react';
-import clsx from 'clsx';
+import React, { useRef, RefObject, CSSProperties } from 'react';
 import type Color from 'color';
 import type { Maybe } from 'true-myth';
 import { doubleFF, singleFF, Flamebearer } from '@pyroscope/models/src';
+// until ui is moved to its own package this should do it
+// eslint-disable-next-line import/no-extraneous-dependencies
+import TableUI, {
+  useTableSort,
+  BodyRow,
+  TableBodyType,
+} from '@webapp/ui/Table';
 import TableTooltip from './Tooltip/TableTooltip';
 import { getFormatter } from './format/format';
 import {
@@ -13,7 +19,6 @@ import {
 import { fitIntoTableCell, FitModes } from './fitMode/fitMode';
 import { isMatch } from './search';
 import type { FlamegraphPalette } from './FlameGraph/FlameGraphComponent/colorPalette';
-import styles from './ProfilerTable.module.scss';
 
 const zero = (v?: number) => v || 0;
 
@@ -34,6 +39,7 @@ interface DoubleCell {
   totalRght: number;
   totalDiff: number;
 }
+type DoubleSortByKeys = keyof DoubleCell & 'name';
 
 function generateCellSingle(
   ff: typeof singleFF,
@@ -163,9 +169,10 @@ const tableFormatSingle: {
   sortable: number;
   name: 'name' | 'self' | 'total';
   label: string;
+  default?: boolean;
 }[] = [
   { sortable: 1, name: 'name', label: 'Location' },
-  { sortable: 1, name: 'self', label: 'Self' },
+  { sortable: 1, name: 'self', label: 'Self', default: true },
   { sortable: 1, name: 'total', label: 'Total' },
 ];
 
@@ -182,13 +189,29 @@ const tableFormatDiffDef: Record<
       | 'totalRght'
       | 'totalDiff';
     label: string;
+    default?: boolean;
   }
 > = {
   name: { sortable: 1, name: 'name', label: 'Location' },
-  selfLeft: { sortable: 1, name: 'selfLeft', label: 'Self (Left)' },
+  selfLeft: {
+    sortable: 1,
+    name: 'selfLeft',
+    label: 'Self (Left)',
+    default: true,
+  },
   selfRght: { sortable: 1, name: 'selfRght', label: 'Self (Right)' },
-  selfDiff: { sortable: 1, name: 'selfDiff', label: 'Self (Diff)' },
-  totalLeft: { sortable: 1, name: 'totalLeft', label: 'Total (Left)' },
+  selfDiff: {
+    sortable: 1,
+    name: 'selfDiff',
+    label: 'Self (Diff)',
+    default: true,
+  },
+  totalLeft: {
+    sortable: 1,
+    name: 'totalLeft',
+    label: 'Total (Left)',
+    default: true,
+  },
   totalRght: { sortable: 1, name: 'totalRght', label: 'Total (Right)' },
   totalDiff: { sortable: 1, name: 'totalDiff', label: 'Total (Diff)' },
 };
@@ -202,9 +225,6 @@ const tableFormatDiff = ((def) => ({
 function Table({
   tableBodyRef,
   flamebearer,
-  updateSortBy,
-  sortBy,
-  sortByDirection,
   viewDiff,
   fitMode,
   handleTableItemClick,
@@ -213,55 +233,40 @@ function Table({
   palette,
 }: ProfilerTableProps) {
   const tableFormat = !viewDiff ? tableFormatSingle : tableFormatDiff[viewDiff];
+  const tableSortProps = useTableSort(tableFormat);
+  const table = {
+    headRow: tableFormat,
+    ...getTableBody({
+      flamebearer,
+      sortBy: tableSortProps.sortBy,
+      sortByDirection: tableSortProps.sortByDirection,
+      viewDiff,
+      fitMode,
+      handleTableItemClick,
+      highlightQuery,
+      palette,
+      selectedItem,
+    }),
+  };
 
   return (
-    <table
-      className={`flamegraph-table ${styles.table}`}
-      data-testid="table-view"
-    >
-      <thead>
-        <tr>
-          {tableFormat.map((v: typeof tableFormat[number], idx: number) =>
-            !v.sortable ? (
-              // eslint-disable-next-line react/no-array-index-key
-              <th key={idx}>{v.label}</th>
-            ) : (
-              <th
-                // eslint-disable-next-line react/no-array-index-key
-                key={idx}
-                className="sortable"
-                onClick={() => updateSortBy(v.name)}
-              >
-                {v.label}
-                <span
-                  className={clsx('sort-arrow', {
-                    [sortByDirection]: sortBy === v.name,
-                  })}
-                />
-              </th>
-            )
-          )}
-        </tr>
-      </thead>
-      <tbody ref={tableBodyRef}>
-        <TableBody
-          flamebearer={flamebearer}
-          sortBy={sortBy}
-          sortByDirection={sortByDirection}
-          viewDiff={viewDiff}
-          fitMode={fitMode}
-          handleTableItemClick={handleTableItemClick}
-          highlightQuery={highlightQuery}
-          palette={palette}
-          selectedItem={selectedItem}
-        />
-      </tbody>
-    </table>
+    <TableUI
+      /* eslint-disable-next-line react/jsx-props-no-spreading */
+      {...tableSortProps}
+      tableBodyRef={tableBodyRef}
+      table={table}
+      className="flamegraph-table"
+    />
   );
 }
 
-// const TableBody = React.memo(
-const TableBody = ({
+interface GetTableBodyRowsProps
+  extends Omit<ProfilerTableProps, 'tableBodyRef'> {
+  sortBy: string;
+  sortByDirection: string;
+}
+
+const getTableBody = ({
   flamebearer,
   sortBy,
   sortByDirection,
@@ -271,40 +276,38 @@ const TableBody = ({
   highlightQuery,
   palette,
   selectedItem,
-}: Omit<ProfilerTableProps, 'updateSortBy' | 'tableBodyRef'>) => {
+}: GetTableBodyRowsProps): TableBodyType => {
   const { numTicks, maxSelf, sampleRate, spyName, units } = flamebearer;
 
-  const table = generateTable(flamebearer).sort((a, b) => b.total - a.total);
-
+  const tableBodyCells = generateTable(flamebearer).sort(
+    (a, b) => b.total - a.total
+  );
   const m = sortByDirection === 'asc' ? 1 : -1;
+  let sorted: typeof tableBodyCells;
 
-  let sorted: typeof table;
   if (sortBy === 'name') {
-    sorted = table.sort((a, b) => m * a[sortBy].localeCompare(b[sortBy]));
+    sorted = tableBodyCells.sort(
+      (a, b) => m * a[sortBy].localeCompare(b[sortBy])
+    );
   } else {
     switch (sortBy) {
       case 'total':
       case 'self': {
-        sorted = table.sort((a, b) => m * (a[sortBy] - b[sortBy]));
+        sorted = tableBodyCells.sort((a, b) => m * (a[sortBy] - b[sortBy]));
         break;
       }
 
       // sorting by all other fields means it must be a double
       default: {
-        sorted = (table as (DoubleCell & { name: string })[]).sort(
-          (a, b) => m * (a[sortBy] - b[sortBy])
+        sorted = (tableBodyCells as (DoubleCell & { name: string })[]).sort(
+          (a, b) =>
+            m * (a[sortBy as DoubleSortByKeys] - b[sortBy as DoubleSortByKeys])
         );
       }
     }
   }
 
-  // The problem is that when you switch apps or time-range and the function
-  //   names stay the same it leads to an issue where rows don't get re-rendered
-  // So we force a rerender each time.
-  const renderID = Math.random();
-
   const formatter = getFormatter(numTicks, sampleRate, units);
-
   const isRowSelected = (name: string) => {
     if (selectedItem.isJust) {
       return name === selectedItem.value;
@@ -313,179 +316,173 @@ const TableBody = ({
     return false;
   };
 
-  const nameCell = (x: { name: string }, style: React.CSSProperties) => (
-    <td>
-      <button className="table-item-button">
-        <span className="color-reference" style={style} />
-        <div className="symbol-name" style={fitIntoTableCell(fitMode)}>
-          {x.name}
-        </div>
-      </button>
-    </td>
+  const nameCell = (x: { name: string }, style: CSSProperties) => (
+    <button className="table-item-button">
+      <span className="color-reference" style={style} />
+      <div className="symbol-name" style={fitIntoTableCell(fitMode)}>
+        {x.name}
+      </div>
+    </button>
   );
 
-  const renderSingleRow = (
+  const getSingleRow = (
     x: SingleCell & { name: string },
     color: Color,
-    style: React.CSSProperties
-  ) => (
-    <tr
-      data-row={`${x.name};${x.self};${x.total};${x.type}`}
-      key={`${x.name}${renderID}`}
-      onClick={() => handleTableItemClick(x)}
-      className={`${isRowSelected(x.name) && styles.rowSelected}`}
-    >
-      {nameCell(x, style)}
-      <td style={backgroundImageStyle(x.self, maxSelf, color)}>
-        {formatter.format(x.self, sampleRate)}
-      </td>
-      <td style={backgroundImageStyle(x.total, numTicks, color)}>
-        {formatter.format(x.total, sampleRate)}
-      </td>
-    </tr>
-  );
+    style: CSSProperties
+  ): BodyRow => ({
+    'data-row': `${x.name};${x.self};${x.total};${x.type}`,
+    isRowSelected: isRowSelected(x.name),
+    onClick: () => handleTableItemClick(x),
+    cells: [
+      { value: nameCell(x, style) },
+      {
+        value: formatter.format(x.self, sampleRate),
+        style: backgroundImageStyle(x.self, maxSelf, color),
+      },
+      {
+        value: formatter.format(x.total, sampleRate),
+        style: backgroundImageStyle(x.total, numTicks, color),
+      },
+    ],
+  });
 
-  const renderDoubleRow = (function () {
+  const getDoubleRow = (() => {
     switch (viewDiff) {
-      case 'self': {
+      case 'self':
         return (
           x: DoubleCell & { name: string },
           color: Color,
-          style: React.CSSProperties
-        ) => (
-          <tr
-            key={`${x.name}${renderID}`}
-            onClick={() => handleTableItemClick(x)}
-            className={`${isRowSelected(x.name) && styles.rowSelected}`}
-          >
-            {nameCell(x, style)}
-            {/* NOTE: it seems React does not understand multiple backgrounds, have to workaround:  */}
-            {/*   The `style` prop expects a mapping from style properties to values, not a string. */}
-            <td
-              style={backgroundImageDiffStyle(
+          style: CSSProperties
+        ): BodyRow & { type: 'self' } => ({
+          type: 'self',
+          isRowSelected: isRowSelected(x.name),
+          onClick: () => handleTableItemClick(x),
+          cells: [
+            { value: nameCell(x, style) },
+            {
+              style: backgroundImageDiffStyle(
                 palette,
                 x.selfLeft,
                 x.selfRght,
                 maxSelf,
                 color,
                 'L'
-              )}
-            >
-              <span title={formatter.format(x.selfLeft, sampleRate)}>
-                {formatter.format(x.selfLeft, sampleRate)}
-              </span>
-            </td>
-            <td
-              style={backgroundImageDiffStyle(
+              ),
+              value: (() => (
+                <span title={formatter.format(x.selfLeft, sampleRate)}>
+                  {formatter.format(x.selfLeft, sampleRate)}
+                </span>
+              ))(),
+            },
+            {
+              style: backgroundImageDiffStyle(
                 palette,
                 x.selfLeft,
                 x.selfRght,
                 maxSelf,
                 color,
                 'R'
-              )}
-            >
-              <span title={formatter.format(x.selfRght, sampleRate)}>
-                {formatter.format(x.selfRght, sampleRate)}
-              </span>
-            </td>
-          </tr>
-        );
-      }
-
-      case 'total': {
+              ),
+              value: (
+                <span title={formatter.format(x.selfRght, sampleRate)}>
+                  {formatter.format(x.selfRght, sampleRate)}
+                </span>
+              ),
+            },
+          ],
+        });
+      case 'total':
         return (
           x: DoubleCell & { name: string },
           color: Color,
-          style: React.CSSProperties
-        ) => (
-          <tr
-            key={`${x.name}${renderID}`}
-            onClick={() => handleTableItemClick(x)}
-            className={`${isRowSelected(x.name) && styles.rowSelected}`}
-          >
-            {nameCell(x, style)}
-            <td
-              style={backgroundImageDiffStyle(
+          style: CSSProperties
+        ): BodyRow & { type: 'total' } => ({
+          type: 'total',
+          isRowSelected: isRowSelected(x.name),
+          onClick: () => handleTableItemClick(x),
+          cells: [
+            { value: nameCell(x, style) },
+            {
+              style: backgroundImageDiffStyle(
                 palette,
                 x.totalLeft,
                 x.totalRght,
                 numTicks / 2,
                 color,
                 'L'
-              )}
-            >
-              <span title={formatter.format(x.totalLeft, sampleRate)}>
-                {formatter.format(x.totalLeft, sampleRate)}
-              </span>
-            </td>
-            <td
-              style={backgroundImageDiffStyle(
+              ),
+              value: (() => (
+                <span title={formatter.format(x.totalLeft, sampleRate)}>
+                  {formatter.format(x.totalLeft, sampleRate)}
+                </span>
+              ))(),
+            },
+            {
+              style: backgroundImageDiffStyle(
                 palette,
                 x.totalLeft,
                 x.totalRght,
                 numTicks / 2,
                 color,
                 'R'
-              )}
-            >
-              <span title={formatter.format(x.totalRght, sampleRate)}>
-                {formatter.format(x.totalRght, sampleRate)}
-              </span>
-            </td>
-          </tr>
-        );
-      }
-
-      case 'diff': {
+              ),
+              value: (() => (
+                <span title={formatter.format(x.totalRght, sampleRate)}>
+                  {formatter.format(x.totalRght, sampleRate)}
+                </span>
+              ))(),
+            },
+          ],
+        });
+      case 'diff':
         return (
           x: DoubleCell & { name: string },
           color: Color,
-          style: React.CSSProperties
-        ) => (
-          <tr
-            key={`${x.name}${renderID}`}
-            onClick={() => handleTableItemClick(x)}
-            className={`${isRowSelected(x.name) && styles.rowSelected}`}
-          >
-            {nameCell(x, style)}
-            <td
-              style={backgroundImageDiffStyle(
+          style: CSSProperties
+        ): BodyRow & { type: 'diff' } => ({
+          type: 'diff',
+          isRowSelected: isRowSelected(x.name),
+          onClick: () => handleTableItemClick(x),
+          cells: [
+            { value: nameCell(x, style) },
+            {
+              style: backgroundImageDiffStyle(
                 palette,
                 x.selfLeft,
                 x.selfRght,
                 maxSelf,
                 defaultColor
-              )}
-            >
-              <span title={formatter.format(x.selfDiff, sampleRate)}>
-                {formatter.format(x.selfDiff, sampleRate)}
-              </span>
-            </td>
-            <td
-              style={backgroundImageDiffStyle(
+              ),
+              value: (() => (
+                <span title={formatter.format(x.selfDiff, sampleRate)}>
+                  {formatter.format(x.selfDiff, sampleRate)}
+                </span>
+              ))(),
+            },
+            {
+              style: backgroundImageDiffStyle(
                 palette,
                 x.totalLeft,
                 x.totalRght,
                 numTicks / 2,
                 color
-              )}
-            >
-              <span title={formatter.format(x.totalDiff, sampleRate)}>
-                {formatter.format(x.totalDiff, sampleRate)}
-              </span>
-            </td>
-          </tr>
-        );
-      }
-
-      default: {
-        return () => <div>Unsupported</div>;
-      }
+              ),
+              value: (() => (
+                <span title={formatter.format(x.totalDiff, sampleRate)}>
+                  {formatter.format(x.totalDiff, sampleRate)}
+                </span>
+              ))(),
+            },
+          ],
+        });
+      default:
+        return (): { type: 'unsupported' } => ({
+          type: 'unsupported',
+        });
     }
   })();
 
-  const items = sorted
+  const rows = sorted
     .filter((x) => {
       if (!highlightQuery) {
         return true;
@@ -493,7 +490,7 @@ const TableBody = ({
 
       return isMatch(highlightQuery, x.name);
     })
-    .map((x) => {
+    .reduce((acc, x) => {
       const pn = getPackageNameFromStackTrace(spyName, x.name);
       const color = viewDiff
         ? defaultColor
@@ -503,29 +500,28 @@ const TableBody = ({
       };
 
       if (x.type === 'double') {
-        return renderDoubleRow(x, color, style);
+        const doubleRow = getDoubleRow(x, color, style);
+
+        if (doubleRow.type === 'unsupported') return acc;
+
+        acc.push(doubleRow);
+        return acc;
       }
 
-      return renderSingleRow(x, color, style);
-    });
+      acc.push(getSingleRow(x, color, style));
+      return acc;
+    }, [] as BodyRow[]);
 
-  return <>{items}</>;
+  return rows.length > 0
+    ? { bodyRows: rows, type: 'filled' as const }
+    : {
+        value: <div className="unsupported-format">Unsupported</div>,
+        type: 'not-filled' as const,
+      };
 };
 
 export interface ProfilerTableProps {
   flamebearer: Flamebearer;
-  sortByDirection: 'asc' | 'desc';
-  sortBy:
-    | 'name'
-    | 'self'
-    | 'total'
-    | 'selfDiff'
-    | 'totalDiff'
-    | 'selfLeft'
-    | 'selfRght'
-    | 'totalLeft'
-    | 'totalRght';
-  updateSortBy: (s: ProfilerTableProps['sortBy']) => void;
   viewDiff?: 'diff' | 'total' | 'self' | false;
   fitMode: FitModes;
   handleTableItemClick: (tableItem: { name: string }) => void;
@@ -538,9 +534,6 @@ export interface ProfilerTableProps {
 
 export default function ProfilerTable({
   flamebearer,
-  sortByDirection,
-  sortBy,
-  updateSortBy,
   viewDiff,
   fitMode,
   handleTableItemClick,
@@ -555,9 +548,6 @@ export default function ProfilerTable({
       <Table
         tableBodyRef={tableBodyRef}
         flamebearer={flamebearer}
-        updateSortBy={updateSortBy}
-        sortBy={sortBy}
-        sortByDirection={sortByDirection}
         viewDiff={viewDiff}
         fitMode={fitMode}
         highlightQuery={highlightQuery}