+++ webapp/javascript/pages/TagExplorerView.tsx
@@ -1,9 +1,8 @@
-import React, { useEffect, useState } from 'react';
-import { NavLink, useLocation, Redirect } from 'react-router-dom';
+import React, { useEffect } from 'react';
+import { NavLink, useLocation } from 'react-router-dom';
 import type { Maybe } from 'true-myth';
 import type { ClickEvent } from '@szhsin/react-menu';
 import Color from 'color';
-import OutsideClickHandler from 'react-outside-click-handler';
 
 import type { Profile } from '@pyroscope/models/src';
 import Box from '@webapp/ui/Box';
@@ -11,11 +10,12 @@ import Toolbar from '@webapp/components/Toolbar';
 import ExportData from '@webapp/components/ExportData';
 import TimelineChartWrapper, {
   TimelineGroupData,
-} from '@webapp/components/TimelineChartWrapper';
-import { FlamegraphRenderer, DefaultPalette } from '@pyroscope/flamegraph/src';
+} from '@webapp/components/TimelineChart/TimelineChartWrapper';
+import { FlamegraphRenderer } from '@pyroscope/flamegraph/src';
 import Dropdown, { MenuItem } from '@webapp/ui/Dropdown';
 import LoadingSpinner from '@webapp/ui/LoadingSpinner';
-import ViewTagsSelectLinkModal from '@webapp/pages/tagExplorer/components/ViewTagsSelectLinkModal';
+import TagsSelector from '@webapp/pages/tagExplorer/components/TagsSelector';
+import TableUI, { BodyRow } from '@webapp/ui/Table';
 import useColorMode from '@webapp/hooks/colorMode.hook';
 import useTimeZone from '@webapp/hooks/timeZone.hook';
 import { appendLabelToQuery } from '@webapp/util/query';
@@ -31,13 +31,50 @@ import {
   TagsState,
   fetchTagExplorerView,
   fetchTagExplorerViewProfile,
-  appWithoutTagsWhereDropdownOptionName,
+  ALL_TAGS,
 } from '@webapp/redux/reducers/continuous';
 import { queryToAppName } from '@webapp/models/query';
+import PageTitle from '@webapp/components/PageTitle';
 import { calculateMean, calculateStdDeviation } from './math';
 import { PAGES } from './constants';
 
+// eslint-disable-next-line css-modules/no-unused-class
 import styles from './TagExplorerView.module.scss';
+import { formatTitle } from './formatTitle';
+
+const TIMELINE_SERIES_COLORS = [
+  Color.rgb(115, 191, 105),
+  Color.rgb(244, 213, 152),
+  Color.rgb(174, 162, 224),
+  Color.rgb(112, 219, 237),
+  Color.rgb(224, 117, 45),
+  Color.rgb(191, 27, 0),
+  Color.rgb(229, 168, 226),
+  Color.rgb(183, 219, 171),
+  Color.rgb(229, 172, 14),
+  Color.rgb(10, 80, 161),
+  Color.rgb(97, 77, 147),
+  Color.rgb(100, 176, 200),
+  Color.rgb(150, 45, 130),
+  Color.rgb(214, 131, 206),
+  Color.rgb(63, 104, 51),
+  Color.rgb(150, 115, 2),
+  Color.rgb(5, 43, 81),
+  Color.rgb(63, 43, 91),
+  Color.rgb(153, 68, 10),
+  Color.rgb(88, 20, 12),
+  Color.rgb(224, 249, 215),
+  Color.rgb(252, 234, 202),
+  Color.rgb(207, 250, 255),
+  Color.rgb(249, 226, 210),
+  Color.rgb(252, 226, 222),
+  Color.rgb(186, 223, 244),
+  Color.rgb(249, 217, 249),
+  Color.rgb(222, 218, 247),
+];
+
+const getTimelineColor = (index: number, palette: Color[]): Color =>
+  Color(palette[index % (palette.length - 1)]);
 
 function TagExplorerView() {
   const { offset } = useTimeZone();
@@ -84,12 +121,10 @@ function TagExplorerView() {
       case 'reloading':
         const groups = Object.entries(tagExplorerView.groups).reduce(
           (acc, [tagName, data], index) => {
-            if (index === 15) return acc;
-
             acc.push({
               tagName,
               data,
-              color: Color(DefaultPalette.colors[index]),
+              color: getTimelineColor(index, TIMELINE_SERIES_COLORS),
             });
 
             return acc;
@@ -119,8 +154,8 @@ function TagExplorerView() {
 
   const { groupsData, activeTagProfile } = getGroupsData();
 
-  const handleGroupByTagValueChange = (groupByTagValue: string) => {
-    dispatch(actions.setTagExplorerViewGroupByTagValue(groupByTagValue));
+  const handleGroupByTagValueChange = (v: string) => {
+    dispatch(actions.setTagExplorerViewGroupByTagValue(v));
   };
 
   const handleGroupedByTagChange = (value: string) => {
@@ -153,93 +188,87 @@ function TagExplorerView() {
   }, [] as string[]);
 
   return (
-    <div className={styles.tagExplorerView} data-testid="tag-explorer-view">
-      <Toolbar hideTagsBar />
-      <Box>
-        <ExploreHeader
-          appName={appName}
-          tags={tags}
-          whereDropdownItems={whereDropdownItems}
-          selectedTag={tagExplorerView.groupByTag}
-          selectedTagValue={tagExplorerView.groupByTagValue}
-          handleGroupByTagChange={handleGroupedByTagChange}
-          handleGroupByTagValueChange={handleGroupByTagValueChange}
-        />
-        <div className={styles.timelineWrapper}>
-          {type === 'loading' ? (
-            <LoadingSpinner />
-          ) : (
-            <TimelineChartWrapper
-              mode="multiple"
-              timezone={offset === 0 ? 'utc' : 'browser'}
-              data-testid="timeline-explore-page"
-              id="timeline-chart-explore-page"
-              timelineGroups={filteredGroupsData}
-              // to not "dim" timelines when "All" option is selected
-              activeGroup={
-                groupByTagValue !== appWithoutTagsWhereDropdownOptionName
-                  ? groupByTagValue
-                  : ''
-              }
-              showTagsLegend={filteredGroupsData.length > 1}
-              handleGroupByTagValueChange={handleGroupByTagValueChange}
-              onSelect={(from, until) =>
-                dispatch(setDateRange({ from, until }))
-              }
-              height="125px"
-              format="lines"
-            />
-          )}
-        </div>
-      </Box>
-      <Box>
-        <Table
-          appName={appName.unwrapOr('')}
-          whereDropdownItems={whereDropdownItems}
-          groupByTag={groupByTag}
-          groupByTagValue={groupByTagValue}
-          groupsData={filteredGroupsData}
-          handleGroupByTagValueChange={handleGroupByTagValueChange}
-          isLoading={type === 'loading'}
-        />
-      </Box>
-      <Box>
-        <div className={styles.flamegraphWrapper}>
-          {type === 'loading' ? (
-            <LoadingSpinner />
-          ) : (
-            <FlamegraphRenderer
-              showCredit={false}
-              profile={activeTagProfile}
-              colorMode={colorMode}
-              ExportData={
-                activeTagProfile && (
-                  <ExportData
-                    flamebearer={activeTagProfile}
-                    exportPNG
-                    exportJSON
-                    exportPprof
-                    exportHTML
-                    exportFlamegraphDotCom
-                    exportFlamegraphDotComFn={exportFlamegraphDotComFn}
-                  />
-                )
-              }
-            />
-          )}
-        </div>
-      </Box>
-    </div>
+    <>
+      <PageTitle title={formatTitle('Tag Explorer View', query)} />
+      <div className={styles.tagExplorerView} data-testid="tag-explorer-view">
+        <Toolbar hideTagsBar />
+        <Box>
+          <ExploreHeader
+            appName={appName}
+            tags={tags}
+            whereDropdownItems={whereDropdownItems}
+            selectedTag={tagExplorerView.groupByTag}
+            selectedTagValue={tagExplorerView.groupByTagValue}
+            handleGroupByTagChange={handleGroupedByTagChange}
+            handleGroupByTagValueChange={handleGroupByTagValueChange}
+          />
+          <div className={styles.timelineWrapper}>
+            {type === 'loading' ? (
+              <LoadingSpinner />
+            ) : (
+              <TimelineChartWrapper
+                selectionType="double"
+                mode="multiple"
+                timezone={offset === 0 ? 'utc' : 'browser'}
+                data-testid="timeline-explore-page"
+                id="timeline-chart-explore-page"
+                timelineGroups={filteredGroupsData}
+                // to not "dim" timelines when "All" option is selected
+                activeGroup={
+                  groupByTagValue !== ALL_TAGS ? groupByTagValue : ''
+                }
+                showTagsLegend={filteredGroupsData.length > 1}
+                handleGroupByTagValueChange={handleGroupByTagValueChange}
+                onSelect={(from, until) =>
+                  dispatch(setDateRange({ from, until }))
+                }
+                height="125px"
+                format="lines"
+              />
+            )}
+          </div>
+        </Box>
+        <Box>
+          <Table
+            appName={appName.unwrapOr('')}
+            whereDropdownItems={whereDropdownItems}
+            groupByTag={groupByTag}
+            groupByTagValue={groupByTagValue}
+            groupsData={filteredGroupsData}
+            handleGroupByTagValueChange={handleGroupByTagValueChange}
+            isLoading={type === 'loading'}
+          />
+        </Box>
+        <Box>
+          <div className={styles.flamegraphWrapper}>
+            {type === 'loading' ? (
+              <LoadingSpinner />
+            ) : (
+              <FlamegraphRenderer
+                showCredit={false}
+                profile={activeTagProfile}
+                colorMode={colorMode}
+                ExportData={
+                  activeTagProfile && (
+                    <ExportData
+                      flamebearer={activeTagProfile}
+                      exportPNG
+                      exportJSON
+                      exportPprof
+                      exportHTML
+                      exportFlamegraphDotCom
+                      exportFlamegraphDotComFn={exportFlamegraphDotComFn}
+                    />
+                  )
+                }
+              />
+            )}
+          </div>
+        </Box>
+      </div>
+    </>
   );
 }
-
-const defaultLinkTagsSelectModalData = {
-  baselineTag: '',
-  comparisonTag: '',
-  linkName: '',
-  isModalOpen: false,
-  shouldRedirect: false,
-};
 
 function Table({
   appName,
@@ -258,13 +287,6 @@ function Table({
   isLoading: boolean;
   handleGroupByTagValueChange: (groupedByTagValue: string) => void;
 }) {
-  const [linkTagsSelectModalData, setLinkTagsSelectModalData] = useState(
-    defaultLinkTagsSelectModalData
-  );
-  const handleOutsideModalClick = () => {
-    setLinkTagsSelectModalData(defaultLinkTagsSelectModalData);
-  };
-
   const { search } = useLocation();
   const isTagSelected = (tag: string) => tag === groupByTagValue;
 
@@ -272,32 +294,12 @@ function Table({
     if (value !== groupByTagValue) {
       handleGroupByTagValueChange(value);
     } else {
-      handleGroupByTagValueChange(appWithoutTagsWhereDropdownOptionName);
+      handleGroupByTagValueChange(ALL_TAGS);
     }
   };
 
-  const handleLinkModalOpen = (linkName: 'Comparison' | 'Diff') => {
-    setLinkTagsSelectModalData((currState) => ({
-      ...currState,
-      isModalOpen: true,
-      linkName,
-    }));
-  };
-
-  if (linkTagsSelectModalData.shouldRedirect) {
-    return (
-      <Redirect
-        to={
-          (linkTagsSelectModalData.linkName === 'Diff'
-            ? PAGES.COMPARISON_DIFF_VIEW
-            : PAGES.COMPARISON_VIEW) + search
-        }
-      />
-    );
-  }
-
   const getSingleViewSearch = () => {
-    if (!groupByTagValue) return search;
+    if (!groupByTagValue || ALL_TAGS) return search;
 
     const searchParams = new URLSearchParams(search);
     searchParams.delete('query');
@@ -306,6 +308,60 @@ function Table({
       appendLabelToQuery(`${appName}{}`, groupByTag, groupByTagValue)
     );
     return `?${searchParams.toString()}`;
+  };
+
+  const headRow = [
+    // when groupByTag is not selected table represents single "application without tags" group
+    {
+      name: 'name',
+      label: `${groupByTag === '' ? 'App' : 'Tag'} name`,
+      sortable: 0,
+    },
+    { name: 'eventCount', label: 'Event count', sortable: 0 },
+    { name: 'avgSamples', label: 'avg samples', sortable: 0 },
+    { name: 'stdDeviation', label: 'std deviation samples', sortable: 0 },
+    { name: 'minSamples', label: 'min samples', sortable: 0 },
+    { name: 'maxSamples', label: 'max samples', sortable: 0 },
+  ];
+
+  const bodyRows = groupsData.reduce(
+    (acc, { tagName, color, data }): BodyRow[] => {
+      const mean = calculateMean(data.samples);
+      const row = {
+        isRowSelected: isTagSelected(tagName),
+        // prevent clicking on single "application without tags" group row
+        onClick:
+          tagName !== appName ? () => handleTableRowClick(tagName) : undefined,
+        cells: [
+          {
+            value: (
+              <div className={styles.tagName}>
+                <span
+                  className={styles.tagColor}
+                  style={{ backgroundColor: color?.toString() }}
+                />
+                {tagName}
+              </div>
+            ),
+          },
+          { value: data.samples.length },
+          { value: mean.toFixed(2) },
+          { value: calculateStdDeviation(data.samples, mean).toFixed(2) },
+          { value: Math.min(...data.samples) },
+          { value: Math.max(...data.samples) },
+        ],
+      };
+      acc.push(row);
+
+      return acc;
+    },
+    [] as BodyRow[]
+  );
+  const table = {
+    headRow,
+    ...(isLoading
+      ? { type: 'not-filled' as const, value: <LoadingSpinner /> }
+      : { type: 'filled' as const, bodyRows }),
   };
 
   return (
@@ -322,88 +378,21 @@ function Table({
           >
             Single
           </NavLink>
-          <button
-            className={styles.buttonName}
-            onClick={() => handleLinkModalOpen('Comparison')}
-          >
-            Comparison
-          </button>
-          <button
-            className={styles.buttonName}
-            onClick={() => handleLinkModalOpen('Diff')}
-          >
-            Diff
-          </button>
-          {linkTagsSelectModalData.isModalOpen && (
-            <OutsideClickHandler onOutsideClick={handleOutsideModalClick}>
-              <ViewTagsSelectLinkModal
-                whereDropdownItems={whereDropdownItems}
-                groupByTag={groupByTag}
-                appName={appName}
-                /* eslint-disable-next-line react/jsx-props-no-spreading */
-                {...linkTagsSelectModalData}
-                setLinkTagsSelectModalData={setLinkTagsSelectModalData}
-              />
-            </OutsideClickHandler>
-          )}
+          <TagsSelector
+            linkName="Comparison"
+            whereDropdownItems={whereDropdownItems}
+            groupByTag={groupByTag}
+            appName={appName}
+          />
+          <TagsSelector
+            linkName="Diff"
+            whereDropdownItems={whereDropdownItems}
+            groupByTag={groupByTag}
+            appName={appName}
+          />
         </div>
       </div>
-      <table className={styles.tagExplorerTable}>
-        <thead>
-          <tr>
-            {/* when groupByTag is not selected table represents single "application without tags" group */}
-            <th>{groupByTag === '' ? 'App' : 'Tag'} name</th>
-            <th>Event count</th>
-            <th>avg samples</th>
-            <th>std deviation samples</th>
-            <th>min samples</th>
-            <th>max samples</th>
-          </tr>
-        </thead>
-        <tbody>
-          {isLoading ? (
-            <tr>
-              <td colSpan={6}>
-                <LoadingSpinner />
-              </td>
-            </tr>
-          ) : (
-            groupsData.map(({ tagName, color, data }) => {
-              const mean = calculateMean(data.samples);
-
-              return (
-                <tr
-                  className={isTagSelected(tagName) ? styles.activeTagRow : ''}
-                  onClick={
-                    // prevent clicking on single "application without tags" group row
-                    tagName !== appName
-                      ? () => handleTableRowClick(tagName)
-                      : undefined
-                  }
-                  key={tagName}
-                >
-                  <td>
-                    <div className={styles.tagName}>
-                      <span
-                        className={styles.tagColor}
-                        style={{ backgroundColor: color?.toString() }}
-                      />
-                      {tagName}
-                    </div>
-                  </td>
-                  <td>{data.samples.length}</td>
-                  <td>{mean.toFixed(2)}</td>
-                  <td>
-                    {calculateStdDeviation(data.samples, mean).toFixed(2)}
-                  </td>
-                  <td>{Math.min(...data.samples)}</td>
-                  <td>{Math.max(...data.samples)}</td>
-                </tr>
-              );
-            })
-          )}
-        </tbody>
-      </table>
+      <TableUI table={table} className={styles.tagExplorerTable} />
     </>
   );
 }
@@ -462,18 +451,16 @@ function ExploreHeader({
         <Dropdown
           label="select where"
           value={`${selectedTag ? `${selectedTag} = ` : selectedTag} ${
-            selectedTagValue || appWithoutTagsWhereDropdownOptionName
+            selectedTagValue || ALL_TAGS
           }`}
           onItemClick={handleGroupByValueClick}
         >
           {/* always show "All" option */}
-          {[appWithoutTagsWhereDropdownOptionName, ...whereDropdownItems].map(
-            (tagGroupName) => (
-              <MenuItem key={tagGroupName} value={tagGroupName}>
-                {tagGroupName}
-              </MenuItem>
-            )
-          )}
+          {[ALL_TAGS, ...whereDropdownItems].map((tagGroupName) => (
+            <MenuItem key={tagGroupName} value={tagGroupName}>
+              {tagGroupName}
+            </MenuItem>
+          ))}
         </Dropdown>
       </div>
     </div>