+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.51](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.50...@pyroscope/panel-plugin@1.3.51) (2022-08-17)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.50](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.49...@pyroscope/panel-plugin@1.3.50) (2022-08-15)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.3.49](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.48...@pyroscope/panel-plugin@1.3.49) (2022-08-11)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin